Validações do que é recebido na API e API de deletar arquivo

+++ src/controllers/bucket/lib/index.js
@@ -1,10 +1,17 @@
 var { s3 } = require('../../../../config');
+var { errorReturn } = require('../../../../lib');
 
 createBucket = ({ bucketName }) => {
     return new Promise((resolve, reject) => {
+        if (!bucketName) {
+            resolve(errorReturn.bucketName)
+            return
+        }
+
         let bucket_params = {
             Bucket: bucketName
         };
+
         s3.createBucket(bucket_params, (err, data) => {
             if (err) {
                 resolve({
@@ -24,9 +31,15 @@ createBucket = ({ bucketName }) => {
 
 deleteBucket = ({ bucketName }) => {
     return new Promise((resolve, reject) => {
+        if (!bucketName) {
+            resolve(errorReturn.bucketName)
+            return
+        }
+
         let bucket_params = {
             Bucket: bucketName
         };
+
         s3.deleteBucket(bucket_params, (err, data) => {
             if (err) {
                 resolve({
@@ -47,6 +60,7 @@ deleteBucket = ({ bucketName }) => {
 listBuckets = () => {
     return new Promise((resolve, reject) => {
         let bucket_params = {};
+        
         s3.listBuckets(bucket_params, (err, data) => {
             if (err) {
                 resolve({
